+++ test_project_postgres/sample/models.py
@@ -1,3 +1,4 @@
+import django
 from django.db import models
 from django.contrib.postgres.fields import ArrayField, JSONField
 
@@ -7,4 +8,4 @@ class TicTacToeBoard(models.Model):
 
 
 class JsonModel(models.Model):
-    value = JSONField()
+    value = JSONField() if django.VERSION < (3, 1) else models.JSONField()