+++ aldjemy/orm.py
@@ -1,11 +1,12 @@
 from sqlalchemy import orm
+from django.apps import apps
 from django.db.models.fields.related import ForeignKey, OneToOneField, ManyToManyField
 from django.db import connections, router
 from django.db.backends import signals
 from django.conf import settings
 
 from .core import get_engine
-from .table import get_all_django_models, generate_tables
+from .table import generate_tables
 
 
 def get_session(alias="default", recreate=False, **kwargs):
@@ -90,6 +91,7 @@ def _extract_model_attrs(metadata, model, sa_models):
                 secondary=sec_table,
                 primaryjoin=(sec_table.c[sec_column] == table.c[model_pk]),
                 secondaryjoin=(sec_table.c[p_sec_column] == p_table.c[p_name]),
+                overlaps=",".join([fk.m2m_field_name(), fk.m2m_reverse_field_name()]),
             )
             if fk.model() != model:
                 backref = None
@@ -109,9 +111,13 @@ def construct_models(metadata):
     if not metadata.tables:
         generate_tables(metadata)
     tables = metadata.tables
-    models = [model for model in get_all_django_models() if not model._meta.proxy]
+    models = [
+        model
+        for model in apps.get_models(include_auto_created=True)
+        if not model._meta.proxy
+    ]
 
-    sa_models_by_django_models = {}
+    sa_models = {}
 
     for model in models:
 
@@ -132,20 +138,20 @@ def construct_models(metadata):
             {"table": table, "alias": router.db_for_read(model)},
         )
 
-        sa_models_by_django_models[model] = sa_model
+        sa_models[model] = sa_model
 
     for model in models:
-        sa_model = sa_models_by_django_models[model]
+        sa_model = sa_models[model]
         table_name = (
             metadata.schema + "." + model._meta.db_table
             if metadata.schema
             else model._meta.db_table
         )
         table = tables[table_name]
-        attrs = _extract_model_attrs(metadata, model, sa_models_by_django_models)
+        attrs = _extract_model_attrs(metadata, model, sa_models)
         orm.mapper(sa_model, table, attrs)
 
-    return sa_models_by_django_models
+    return sa_models
 
 
 class BaseSQLAModel: