+++ HISTORY.rst
@@ -11,7 +11,17 @@ Incompatible changes:
 * Removed ``aldjemy.core.Cache.sa_models``.
 * Removed ``aldjemy.core.Cache.meta``.
 * Removed ``aldjemy.orm.prepare_models``.
+* Removed ``aldjemy.table.get_all_django_models``.
+* Merged ``aldjemy.types`` and ``aldjemy.postgres`` into ``aldjemy.table``.
 
 Deprecations:
 
 * Deprecated ``aldjemy.core.Cache`` without a warning or replacement.
+
+Fixes:
+
+* Silence some warnings from SQLAlchemy 1.4.
+  Many to many fields create duplicated active relationships,
+  which SQLAlchemy discourages.
+  However, this retains backward compatibility
+  and seems like a reasonable compromise for translating Django models.