+++ HISTORY.rst
@@ -5,6 +5,7 @@ Incompatible changes:
 
 * Dropped support for Python < 3.6.
 * Dropped support for Django < 2.2.
+* Dropped support for SQLALchemy < 1.4.
 * Removed ``aldjemy.to_list``.
 * Removed ``aldjemy.core.get_meta``.
 * Removed ``aldjemy.core.Cache.models``.